Proof of concept linking against Xcproj framework from Swift
The XcprojTest playground runs, but fails with this error:

    error: Playground execution aborted: Execution was interrupted, reason: shared-library-event.

This is promising, but indicates that perhaps dynamic loading won't work
in playgrounds. Next step is to attempt running the same code from a
compiled Swift binary rather than a playground.
