+++ DevToolsCore/DevToolsCore.h
@@ -1,9 +1,10 @@
-#import <DevToolsCore/PBXBuildFile.h>
-#import <DevToolsCore/PBXBuildPhase.h>
-#import <DevToolsCore/PBXContainer.h>
-#import <DevToolsCore/PBXGroup.h>
-#import <DevToolsCore/PBXProject.h>
-#import <DevToolsCore/PBXReference.h>
-#import <DevToolsCore/PBXTarget.h>
-#import <DevToolsCore/XCBuildConfiguration.h>
-#import <DevToolsCore/XCConfigurationList.h>
+#import <Foundation/Foundation.h>
+#import "PBXBuildFile.h"
+#import "PBXBuildPhase.h"
+#import "PBXContainer.h"
+#import "PBXGroup.h"
+#import "PBXProject.h"
+#import "PBXReference.h"
+#import "PBXTarget.h"
+#import "XCBuildConfiguration.h"
+#import "XCConfigurationList.h"
