+++ XcprojTest.playground/Contents.swift
@@ -1,0 +1,7 @@
+import Xcproj
+
+do {
+  try Xcproj.loadFrameworks()
+} catch {
+  print(error)
+}
