+++ xcproj.xcworkspace/contents.xcworkspacedata
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Workspace
+   version = "1.0">
+   <FileRef
+      location = "container:xcproj.xcodeproj">
+   </FileRef>
+   <FileRef
+      location = "group:/Users/adsharp/src/0xced/xcproj/XcprojTest.playground">
+   </FileRef>
+</Workspace>
